Splitting glue code from API
