+++ dt_sim_api/vectorizer/prep_service_model.sh
