+++ dt_sim_api/vectorizer/run_service_model.sh
